requirements.txt
PyMuPDF
google-generativeai
chromadb
scikit-learn
nltk
flask